--- src/screens/communityMemberProfile/CommunityMemberProfile.jsx
@@ -1,58 +1,0 @@
-import React, { useEffect } from 'react'
-import DashboardLayout from '../../layout/dashboardLayout/DashboardLayout'
-import { useParams } from 'react-router-dom'
-import {
-  PersonalInformation,
-  ContactInformation,
-  AdditionalInformation
-} from '../../components/profileFormCard/ProfileFormCard'
-import './CommunityMemberProfile.scss'
-import BackButton from '../../components/backButton/BackButton'
-import EditInformation from '../../components/editInformation/EditInformation'
-import { useDispatch, useSelector } from 'react-redux'
-import { getUserDetails } from '../../actions/userAction'
-import { useHistory } from 'react-router-dom/cjs/react-router-dom.min'
-
-function CommunityMembersProfile () {
-  const { id } = useParams()
-  const dispatch = useDispatch()
-  const history = useHistory()
-  const userDetails = useSelector((state) => state.userDetails)
-  const { user } = userDetails
-  const userLogin = useSelector((state) => state.userLogin)
-  const { userInfo } = userLogin
-
-  useEffect(() => {
-    if (userInfo) {
-      dispatch(getUserDetails(id))
-    }
-  }, [dispatch, history, id, userInfo])
-
-  const editUserInformation = () => {
-    history.push({ pathname: '/register-complete', state: { editInformations: true, user } })
-  }
-
-  return (
-    <DashboardLayout title={user?.firstName + ' ' + user?.lastName}>
-      <div className='x10-4-0-my-personals'>
-        <div className='flex-col-2'>
-          <div className='frame-2923'>
-            <BackButton onClick={() => history.goBack()} />
-          </div>
-          <div className='profile border-1px-onyx'>
-            <div className='profile-info'>
-              {user && <>
-                <PersonalInformation user={user} />
-                {user?.phone === undefined || <ContactInformation user={user} />}
-                <AdditionalInformation user={user} />
-              </>}
-            </div>
-            <EditInformation clickHandler={editUserInformation} image={user?.attachments} follow='follow' />
-          </div>
-        </div>
-      </div>
-    </DashboardLayout>
-  )
-}
-
-export default CommunityMembersProfile
